+++ app/src/main/AndroidManifest.xml
@@ -14,7 +14,7 @@
         android:theme="@style/Theme.HappyBirthday"
         tools:targetApi="31">
         <activity
-            android:name=".MainActivity"
+            android:name=".unit1.UnitOneActivity"
             android:exported="true">
             <intent-filter>
                 <action android:name="android.intent.action.MAIN" />
