+++ app/src/main/java/com/example/happybirthday/unit1/UnitOneActivity.kt
@@ -1,13 +1,13 @@
-package com.example.happybirthday
+package com.example.happybirthday.unit1
 
 import androidx.appcompat.app.AppCompatActivity
 import android.os.Bundle
 import android.widget.Button
 import android.widget.ImageView
 import android.widget.TextView
+import com.example.happybirthday.R
 
-
-class MainActivity : AppCompatActivity() {
+class UnitOneActivity : AppCompatActivity() {
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
         setContentView(R.layout.activity_main)
